service/requirements.txt
numpy==1.24.2
fastapi==0.92.0
uvicorn==0.20.0
requests==2.28.1
onnxruntime==1.14.0
pillow==9.4.0
python-multipart==0.0.5